+++ tests/unit/services/user-key-service.test.ts
@@ -39,57 +39,8 @@ vi.mock("@/lib/services/api-key-quota-tracker", () => ({
 }));
 
 vi.mock("@/lib/db", async () => {
-  const { createClient } = await import("@libsql/client");
-  const { drizzle } = await import("drizzle-orm/libsql");
-  const { is } = await import("drizzle-orm");
-  const { SQLiteTable, getTableConfig } = await import("drizzle-orm/sqlite-core");
-  const fs = await import("fs");
-  const pathMod = await import("path");
-  const schema = await import("@/lib/db/schema-sqlite");
-
-  // Shared-cache in-memory URI so every connection sees the same store.
-  const client = createClient({ url: "file::memory:?cache=shared" });
-
-  // Apply the drizzle-sqlite migrations statement by statement (drizzle's own
-  // libsql migrator chokes on the empty fragments left by breakpoint splitting).
-  const dir = pathMod.resolve(process.cwd(), "drizzle-sqlite");
-  const files = fs
-    .readdirSync(dir)
-    .filter((file) => file.endsWith(".sql"))
-    .sort();
-  for (const file of files) {
-    const raw = fs.readFileSync(pathMod.join(dir, file), "utf8");
-    const statements = raw
-      .split("--> statement-breakpoint")
-      .map((s) => s.trim())
-      .filter((s) => s.length > 0);
-    for (const stmt of statements) {
-      await client.execute(stmt);
-    }
-  }
-
-  // Reconcile migrated tables to schema-sqlite by adding drifted columns.
-  for (const exported of Object.values(schema)) {
-    if (!is(exported, SQLiteTable)) {
-      continue;
-    }
-    const cfg = getTableConfig(exported);
-    const info = await client.execute(`PRAGMA table_info(\`${cfg.name}\`)`);
-    if (info.rows.length === 0) {
-      continue;
-    }
-    const existing = new Set(info.rows.map((row) => String(row.name)));
-    for (const column of cfg.columns) {
-      if (!existing.has(column.name)) {
-        await client.execute(
-          `ALTER TABLE \`${cfg.name}\` ADD COLUMN \`${column.name}\` ${column.getSQLType()}`
-        );
-      }
-    }
-  }
-
-  const db = drizzle(client, { schema });
-  return { db, ...schema };
+  const { createLibsqlMemoryDbModule } = await import("../../helpers/libsql-memory-db");
+  return createLibsqlMemoryDbModule();
 });
 
 import { eq } from "drizzle-orm";
