+++ tests/unit/services/user-service.test.ts
@@ -27,69 +27,8 @@ vi.mock("@/lib/utils/auth", async (importActual) => {
 });
 
 vi.mock("@/lib/db", async () => {
-  const { createClient } = await import("@libsql/client");
-  const { drizzle } = await import("drizzle-orm/libsql");
-  const { is } = await import("drizzle-orm");
-  const { SQLiteTable, getTableConfig } = await import("drizzle-orm/sqlite-core");
-  const fs = await import("fs");
-  const pathMod = await import("path");
-  const schema = await import("@/lib/db/schema-sqlite");
-
-  // A bare ":memory:" libsql database is scoped to a single connection, so the
-  // migration connection and drizzle's query connection would see different
-  // empty databases. A shared-cache in-memory URI gives every connection in
-  // this process the same backing store while staying file-free.
-  const client = createClient({ url: "file::memory:?cache=shared" });
-
-  // Apply the drizzle-sqlite migrations statement by statement. drizzle's libsql
-  // migrator instead batches every fragment including the empty pieces left by
-  // `--> statement-breakpoint` splitting, which makes libsql raise a spurious
-  // "not an error"; splitting and dropping empties ourselves avoids that.
-  const dir = pathMod.resolve(process.cwd(), "drizzle-sqlite");
-  const files = fs
-    .readdirSync(dir)
-    .filter((file) => file.endsWith(".sql"))
-    .sort();
-  for (const file of files) {
-    const raw = fs.readFileSync(pathMod.join(dir, file), "utf8");
-    const statements = raw
-      .split("--> statement-breakpoint")
-      .map((s) => s.trim())
-      .filter((s) => s.length > 0);
-    for (const stmt of statements) {
-      await client.execute(stmt);
-    }
-  }
-
-  // The drizzle-sqlite migrations have drifted from schema-sqlite: some columns
-  // (for example upstreams.official_website_url) are recorded in the snapshot
-  // metadata but no .sql migration ever adds them, and some tables in the schema
-  // are never created by any .sql migration at all. The user-management tests
-  // only touch users / api_keys / upstreams / user_upstreams, which the
-  // migrations do create. Reconcile every migrated table to schema-sqlite by
-  // adding any missing columns (added nullable, which is enough to insert and
-  // read rows in tests), and skip tables the migrations never created.
-  for (const exported of Object.values(schema)) {
-    if (!is(exported, SQLiteTable)) {
-      continue;
-    }
-    const cfg = getTableConfig(exported);
-    const info = await client.execute(`PRAGMA table_info(\`${cfg.name}\`)`);
-    if (info.rows.length === 0) {
-      continue;
-    }
-    const existing = new Set(info.rows.map((row) => String(row.name)));
-    for (const column of cfg.columns) {
-      if (!existing.has(column.name)) {
-        await client.execute(
-          `ALTER TABLE \`${cfg.name}\` ADD COLUMN \`${column.name}\` ${column.getSQLType()}`
-        );
-      }
-    }
-  }
-
-  const db = drizzle(client, { schema });
-  return { db, ...schema };
+  const { createLibsqlMemoryDbModule } = await import("../../helpers/libsql-memory-db");
+  return createLibsqlMemoryDbModule();
 });
 
 import { eq } from "drizzle-orm";
